oct update for favorites page
