+++ app/src/main/AndroidManifest.xml
@@ -12,6 +12,7 @@
 
         <activity android:name=".oct.OCTranspoBusRouteActivity" />
         <activity android:name=".oct.StationDetail" />
+        <activity android:name=".oct.Favourites" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
