+++ app/src/main/res/layout/faourites_list.xml
@@ -6,7 +6,7 @@
     android:layout_height="wrap_content">
 
     <TextView
-        android:id="@+id/historylist"
+        android:id="@+id/favouritesList"
         android:layout_width="289dp"
         android:layout_height="37dp"
         android:layout_marginStart="32dp"
@@ -17,6 +17,7 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <ImageView
+        android:id="@+id/deleteImg"
         android:layout_width="34dp"
         android:layout_height="29dp"
         android:layout_marginStart="280dp"
